[AI-L4] refactor: move bot modules under src
